requirements.txt
flask
flask-cors
pyyaml
telnetlib3
paramiko
redis
requests